% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEASNPtest.R
\name{GSEASNPtest}
\alias{GSEASNPtest}
\title{GSEASNP test implementation}
\usage{
GSEASNPtest(
  data,
  pathways = NA,
  adjust.method = "BH",
  permutations = 10000,
  cores = 2
)
}
\arguments{
\item{data}{input dataframe with column names: "snp", "entrez", "p.value"}

\item{pathways}{object containing a list of GENES and a list of MODULES}

\item{adjust.method}{multiple testing correction method}

\item{permutations}{number of gene set permutations}

\item{cores}{number of cores; default is 2}
}
\value{
A data frame with module names, calculated p-value, and additional statistics.
}
\description{
Performs an GSEASNP test on SNP p-values
}
\examples{
GSEASNPtest(data = example_dataset, pathways = pathway_library, adjust.method = "BH", 
permutations = 10000)

}
