% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAtest.R
\name{GSEAtest}
\alias{GSEAtest}
\title{GSEAtest test implementation}
\usage{
GSEAtest(
  data,
  pathways = NA,
  int.method = "min",
  adjust.int = FALSE,
  int.cor.LD = FALSE,
  adjust.method = "BH",
  permutations = 1000,
  permutation.method = "entrez",
  cores = 2
)
}
\arguments{
\item{data}{input dataframe with column names: "snp", "entrez", "p.value"}

\item{pathways}{object containing a list of GENES and a list of MODULES}

\item{int.method}{integration method (either "min", "fisher', or "stouffer")}

\item{adjust.int}{flag for multiple testing correction in integration}

\item{int.cor.LD}{flag for LD correction in integration}

\item{adjust.method}{multiple testing correction method}

\item{permutations}{number of gene set permutations}

\item{permutation.method}{permutation method (either "entrez", "snp", or "snp_raw")}

\item{cores}{number of cores; if cores > 2 use parallel computing}
}
\value{
A data frame with module names, calculated p-value, and additional statistics.
}
\description{
Performs an GSEAtest test on SNP p-values
}
\examples{
GSEAtest(data = example_dataset, pathways = pathway_library, int.method = "min", 
adjust.int = FALSE, int.cor.LD = FALSE, adjust.method = "BH", permutations = 10000, 
permutation.method = "entrez", cores = 2)

}
