% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAGENTAtest.R
\name{MAGENTAtest}
\alias{MAGENTAtest}
\title{MAGENTA test implementation}
\usage{
MAGENTAtest(
  data,
  pathways = NA,
  int.method = "min",
  adjust.int = FALSE,
  int.cor.LD = FALSE,
  cutoff = 0.05,
  adjust.method = "BH",
  permutations = 1000
)
}
\arguments{
\item{data}{input dataframe with column names: "snp", "entrez", "p.value"}

\item{pathways}{object containing a list of GENES and a list of MODULES}

\item{int.method}{integration method (either "min", "fisher', or "stouffer")}

\item{adjust.int}{flag for multiple testing correction in integration}

\item{int.cor.LD}{flag for LD correction in integration}

\item{cutoff}{enrichment p-value cutoff}

\item{adjust.method}{multiple testing correction method}

\item{permutations}{number of gene set permutations}
}
\value{
A data frame with module names, calculated p-value, and additional statistics.
}
\description{
Performs a MAGENTA test on SNP p-values
}
\examples{
MAGENTAtest(data = example_dataset, pathways = pathway_library, int.method = "min", 
adjust.int = FALSE, int.cor.LD = FALSE, cutoff = 0.05, adjust.method = "BH", permutations = 1000)

}
