% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORAtest.R
\name{ORAtest}
\alias{ORAtest}
\title{ORA test implementation}
\usage{
ORAtest(
  data,
  pathways = NA,
  int.method = "min",
  adjust.int = FALSE,
  int.cor.LD = FALSE,
  adjust.method = "BH"
)
}
\arguments{
\item{data}{input dataframe with column names: "snp", "entrez", "p.value"}

\item{pathways}{object containing a list of GENES and a list of MODULES}

\item{int.method}{integration method (either "min", "fisher', or "stouffer")}

\item{adjust.int}{flag for multiple testing correction in integration}

\item{int.cor.LD}{flag for LD correction in integration}

\item{adjust.method}{multiple testing correction method}
}
\value{
A data frame with module names, calculated p-value, and additional statistics.
}
\description{
Performs a ORA test on SNP p-values
}
\examples{
ORAtest(data = example_dataset, pathways = pathway_library, int.method = "min", 
adjust.int = FALSE, int.cor.LD = TRUE, adjust.method = "BH")

}
